Refactor: app/login
